requirements.txt
aiofiles==23.2.1
annotated-types==0.7.0
anyio==4.9.0
audioop-lts==0.2.1
certifi==2025.1.31
charset-normalizer==3.4.1
click==8.1.8
colorama==0.4.6
fastapi==0.115.12
ffmpeg-python==0.2.0
ffmpy==0.5.0
filelock==3.18.0
fsspec==2025.3.2
future==1.0.0
gradio==5.23.3
gradio_client==1.8.0
groovy==0.1.2
h11==0.14.0
httpcore==1.0.7
httpx==0.28.1
huggingface-hub==0.30.1
idna==3.10
Jinja2==3.1.6
llvmlite==0.44.0
markdown-it-py==3.0.0
MarkupSafe==3.0.2
mdurl==0.1.2
more-itertools==10.6.0
mpmath==1.3.0
networkx==3.4.2
numba==0.61.0
numpy==2.1.3
openai-whisper @ git+https://github.com/openai/whisper.git@517a43ecd132a2089d85f4ebc044728a71d49f6e
orjson==3.10.16
packaging==24.2
pandas==2.2.3
pillow==11.1.0
pydantic==2.11.2
pydantic_core==2.33.1
pydub==0.25.1
Pygments==2.19.1
python-dateutil==2.9.0.post0
python-multipart==0.0.20
pytz==2025.2
PyYAML==6.0.2
regex==2024.11.6
requests==2.32.3
rich==14.0.0
ruff==0.11.3
safehttpx==0.1.6
semantic-version==2.10.0
setuptools==78.1.0
shellingham==1.5.4
six==1.17.0
sniffio==1.3.1
starlette==0.46.1
sympy==1.13.1
tiktoken==0.9.0
tomlkit==0.13.2
torch==2.6.0
torchaudio==2.6.0
torchvision==0.21.0
tqdm==4.67.1
typer==0.15.2
typing-inspection==0.4.0
typing_extensions==4.13.1
tzdata==2025.2
urllib3==2.3.0
uvicorn==0.34.0
websockets==15.0.1
wheel==0.45.1
